+++ src/nti/app/pyramid_zope/i18n/interfaces.py
@@ -6,12 +6,16 @@ I18N related interfaces.
 .. $Id$
 """
 
-from __future__ import print_function, absolute_import, division
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from pyramid.interfaces import IRequest
+
 __docformat__ = "restructuredtext en"
 
 logger = __import__('logging').getLogger(__name__)
 
-from pyramid.interfaces import IRequest
 
 
 class IPreferredLanguagesRequest(IRequest):