+++ src/nti/app/pyramid_zope/i18n/subscribers.py
@@ -1,35 +1,43 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """
 I18N related subscribers.
 
-.. $Id$
 """
 
-from __future__ import print_function, absolute_import, division
-__docformat__ = "restructuredtext en"
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 
-logger = __import__('logging').getLogger(__name__)
+from pyramid.i18n import default_locale_negotiator
+from pyramid.interfaces import IContextFound
 
 from zope import component
 from zope import interface
 
 from zope.i18n.interfaces import IUserPreferredLanguages
+from zope.security.interfaces import IPrincipal
+from zope.authentication.interfaces import IUnauthenticatedPrincipal
 
-from pyramid.i18n import default_locale_negotiator
+from .interfaces import IPreferredLanguagesRequest
 
-from pyramid.interfaces import IContextFound
+__docformat__ = "restructuredtext en"
 
-from nti.app.authentication import get_remote_user
+__all__ = [
+    'adjust_request_interface_for_preferred_languages',
+]
 
-from nti.app.i18n.interfaces import IPreferredLanguagesRequest
+
 
 
 @component.adapter(IContextFound)
-def _adjust_request_interface_for_preferred_languages(event):
+def adjust_request_interface_for_preferred_languages(event):
     """
-    Checks the conditions outlined in this package's documentation
-    and adds a marker interface to the request if they hold true.
+    Checks the conditions outlined in this package's documentation and
+    adds a marker interface (:class:`.IPreferredLanguagesRequest`) to
+    the request if they hold true.
+
+    This is registered as a subscriber for Pyramid's
+    :class:`.IContextFound` event by this package's ``configure.zcml``
     """
     request = event.request
     # Does pyramid's default negotiator, which uses explicit settings
@@ -38,6 +46,7 @@ def _adjust_request_interface_for_preferred_languages(event):
     if default_locale_negotiator(request):
         interface.alsoProvides(request, IPreferredLanguagesRequest)
         return
+
     # What about the zope/plone cookie?
     if request.cookies.get('I18N_LANGUAGE'):
         # For benefit of the default localization machinery
@@ -45,24 +54,11 @@ def _adjust_request_interface_for_preferred_languages(event):
         request._LOCALE_ = request.cookies.get('I18N_LANGUAGE')
         interface.alsoProvides(request, IPreferredLanguagesRequest)
         return
+
     # Ok, is there an authenticated user with preferred languages?
     # (We leave detecting defaults up to the actual policy)
-    try:
-        remote_user = get_remote_user(request)
-    except LookupError:
-        # If we're not in a site, we would get an
-        # InappropriateSiteError here.
-        # We've only actually seen this using pyramid_debugtoolbar
-        # when handling an earlier exception of some type.
-        # See also root_resource_factory...note that that place
-        # specifically checks for the debug toolbar URL, with
-        # request.path.startswith( '/_debug_toolbar/' ),
-        # but I don't really feel like that's a necessary safety check
-        # here
-        remote_user = None
-
-    remote_user_langs = remote_user is not None \
-                    and IUserPreferredLanguages(remote_user)
-    if remote_user_langs and remote_user_langs.getPreferredLanguages():
-        interface.alsoProvides(request, IPreferredLanguagesRequest)
-        return
+    remote_user = IPrincipal(request, None)
+    if remote_user and not IUnauthenticatedPrincipal.providedBy(remote_user):
+        remote_user_langs = IUserPreferredLanguages(remote_user)
+        if remote_user_langs and remote_user_langs.getPreferredLanguages(): # pylint:disable=too-many-function-args
+            interface.alsoProvides(request, IPreferredLanguagesRequest)